+++ compiler/rustc_ast_ir/Cargo.toml
@@ -8,12 +8,14 @@ edition = "2024"
 rustc_data_structures = { path = "../rustc_data_structures", optional = true }
 rustc_macros = { path = "../rustc_macros", optional = true }
 rustc_serialize = { path = "../rustc_serialize", optional = true }
+rustc_span = { path = "../rustc_span", optional = true }
 # tidy-alphabetical-end
 
 [features]
 default = ["nightly"]
 nightly = [
     "dep:rustc_serialize",
+    "dep:rustc_span",
     "dep:rustc_data_structures",
     "dep:rustc_macros",
 ]
